mp-examples: Fix cachetest
Until now I was testing the wrong thing :).

The actual goal was to benchmark the number of increments over an array
whose size changes over time. The procedure is straight forward -
allocate a gigantic array (1GB), increment each element of the first
COUNT element of it 4096 times. At each new procedure call the COUNT is
modified (the steps are 4K).

Signed-off-by: Kristiyan Stoimenov <kristoimenov@gmail.com>